:building_construction: | Changing export way from components
